requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
pytest
httpx